B operations service\  if !operation return operation no exist

+++ backend/components/operations/services/index.js
@@ -48,13 +48,16 @@ module.exports = {
   async deleteOperations(id) {
     try {
       const operation = await operationsRepo.findOperations(id);
-      const amountSign = signAmount(operation.dataValues.amount, operation.dataValues.type);
-      const totalBalance = await operationsRepo.findTotalBalance(userId);
-      const newAmount = totalBalance - amountSign;
-      await operationsRepo.updateTotalBalance(userId, newAmount);
+      if (operation) {
+        const amountSign = signAmount(operation.dataValues.amount, operation.dataValues.type);
+        const totalBalance = await operationsRepo.findTotalBalance(userId);
+        const newAmount = totalBalance - amountSign;
+        await operationsRepo.updateTotalBalance(userId, newAmount);
 
-      const response = await operationsRepo.deleteOperations(id);
-      return (response);
+        const response = await operationsRepo.deleteOperations(id);
+        return (response);
+      }
+      return ({ message: `Not exist operation with id:${id}` });
     } catch (err) {
       return (err);
     }
